Add rule consequene floatplacement=smart
This adds a new option 'smart' for the rule consequence floatplacement.
It tries to place the floating window with as little overlap to other
floating windows as possible. Beside documentation, this also adds
test cases.

Smart window placement is quite hard to test, so I have two test cases
that actually verify the window positions, and one test case that only
adds a bunch of clients without verifying anything (not contributing to
code coverage).