Allow to split non-leaf frames
This extends the 'split' command by an optional argument containing a
frame index. The FrameTree::splitCommand() was quite specific to
FrameLeafs, so I had to change the type of 'frame' and tried to keep the
meaning of the other variables as good as possible for the case where
'frame' is a FrameSplit.

The completion of the 'split' command is still defined in command.cpp.
Migrating this is in the scope of another command.

This implements #906.