Basic implementation of frame objects (#974)
* Basic implementation of frame objects

This commit introduces the long awaited frame objects. For the moment,
the data of frame objects is only read-only, because I have only
DynAttribute_<> access to all the frame objects.

This commit does not introduce no test cases, but instead adds new
assertions to the existing test cases that modify the tiling layout.

So far, the frame objects can be accessed as "TAG_OBJECT.tiling.root".
In the future, there might be a "TAG_OBJECT.floating" and a
"TAG_OBJECT.tiling.focus", hence the two child objects. (Moreover it
fits more easily with the actual implementation in the hlwm source
code).

There is no documentation yet, I think it should be added in a separate.

* Remove unused variable, add whitespace

* Run 'split' later