Frame index 'p' for accessing parent frame (#984)
This extends the frame index possibilities by a new character 'p' that
goes back to the parent frame. In order to avoid a mess with 'special
cases' in the frame index section in the man page, I have reorganized it
a bit improving readability.

There is no real way to test frame indices on their own, but I'm using
the 'dump' command which accepts a frame index parameter, in order to
verify that 'dump' prints the correct subtree.

This implements a feature request raised in the discussion of #906 by
elvisjauld.