Fix crash in smart placement on invisible tags
If multiple floating clients are placed smartly on a tag, then the
'last_scheme' member of Decoration is used to compute the correct
geometry of a client. If this involves clients that have never been
shown, then the 'last_scheme' is not yet initialized.

This commit checks the 'last_scheme' pointer explicitly in
Decoration::inner_to_outer() and adds some test cases which reproduced
the crash earlier. Moreover, I've also added the check in
Decoration::redrawPixmap(), just to be on the safe side.

There is no entry in the NEWS section for this, because this bug was not
present in the last released version.