New constant string format specifier for sprintf
This introduces a new format specifier '%c' in the sprintf format
strings. Of course it never makes sense to manually write something like

    hc sprintf S '%c.name' 'tags.0'

but it makes sense in combination with the other commands that replace
certain identfiers in commands, especiall in comination with the foreach
command.

This change required an update in the completion code for the 'sprintf'
command, and while at it, I fixed a minor bug in the completion. The bug
was that completing the command

   hc sprintf S '%s'

erroneously yielded S.