Add the 'foreach' command for object iteration (#993)
The foreach command iterates over all children of a given object. With
this, it becomes easy to iterate over all clients, tags, monitors. It's
probably only a first version of the 'foreach' because often one needs
to filter the children in a certain way (e.g. 'tags' has more children
than tags because there is something like tags.focus and tags.by-name).

This implements #26.