Add -h and --help to herbstluftwm (#1002)
This is long overdue, but herbstluftwm has been lacking the -h and
--help command line flags. On -h and --help, herbstluftwm now prints a
small explanation what it is, prints all command line options and
mentions 'herbstclient' and the man page. I think this is enough for the
time being, because one never interacts manually with the herbstluftwm
binary (only when calling it mistakenly instead of 'herbstclient'
maybe).