Add per-client keyboard layout script (#1005)
Hello and thank you for the HLWM!

What I was missing when switched from Gnome is the feature to keep the keyboard layout bound to windows individually. Thanks to Herbstluftwm flexibility, this can be recovered. I’ve been testing the script for a week or so on a daily basis, works fine, found no problems.