www: Abandon frame and poisonous colors (#1013)
This updates the visual layout of the web page in order to make it look
more modern:

  - change poisonous colors
  - remove the unnecessary frame around the text content such that the
    text margin on the left and right more look like the margin around
    the text on a piece of paper.

In recent weeks, many spam mails reached the (moderation queue of the)
mailinglist critizing that our web page looks much older and less modern
than those of our competitors; and actually that's true. The present
change tries to fix this.