Fix theme of urgent clients with smart_window (#1017)
As #937 reported, if smart_window_surroundings is enabled, the only
client on a tag correctly uses the minimal theme for surroundings. If
however the client gets an urgency hint, it suddenly uses the normal
theme for displaying the urgent border.

Add a test for this and fix it by caching the most recently used theme
type in order to choose the correct one when drawing the urgent border
(as suggested by Thorsten).