Register theme object children via templates
This introduces a variation of the Child_<> template called
ChildMember_<> which augments a usual C++ member variable by the
awareness of hlwm's object tree. So e.g. in the DecTriple class, I can
change the types of the members 'normal', 'active', 'urgent' from
DecorationScheme to ChildMember_<DecorationScheme> without changing the
rest of the codebase. This has two advantages:

   - It is clear in the class definition, which members are exposed to
     the object tree and which are not
   - It is easily parsable for doc/gendoc.py, which now fully covers the
     theme objects (see the removed exceptions in the test cases)