Fix the value of WM_STATE for hidden windows
A window that is managed but hidden must have WM_STATE=IconicState
whereas hlwm set it to WithdrawnState before. This is described in
ICCCM[1] but also in the implementation note on virtual desktops in
EWMH[2].

Also, when a window is managed but moved to a hidden tag by the rules,
hlwm didn't set WM_STATE at all.

To make it clear to taskbars and pagers that this IconicState does not
mean 'minimized to taskbar', I've added the _NET_WM_STATE_HIDDEN to
_NET_SUPPORTED to make it clear that hlwm would set _NET_WM_STATE_HIDDEN
flag if a window was minimized.

Both these bugs regarding WM_STATE are fixed by the present commit and
checked via test cases.

[1] https://tronche.com/gui/x/icccm/sec-4.html#s-4.1.3.1
[2] https://specifications.freedesktop.org/wm-spec/wm-spec-latest.html#idm46035376813792