Add util to post json-doc diff to github comment
The script `ci/diff-json-doc.py` creates a diff of the output of the
gendoc.py script, and optionally posts this diff as a comment in a PR
discussion. Hence, we directly see how a PR affects the object system.

We should consider adding this to the travis-ci. I found the article
https://damien.pobel.fr/post/github-api-from-travisci/ very helpful
while writing the diff-json-doc.py script.