Make index attribute of tags writable
This requires the respective array operations in the IndexingObject and
also it requires that the EWMH properties are updated accordingly.
While writing test cases for that, I noticed that the EWMH properties
are not updated properly when a tag is removed (this changes the index
of the clients on later tags).

The wiring of the signals now solves both issues.