Add dynamic children DynChild_<> for objects (#1021)
This allows that a child entry of an object is not an explicit Object*
but a lambda instead that dynamically returns the Object* in this child
entry. This is usefull for entries that are indirectly determined by
other values. For those entries it would be much more coding effort to
keep the links up to date, so instead we just look the information up
when the information is queried, just as in the case of DynAttribute_.