Implement client minimization
This adds a new attribute 'minimized' that hides a client until it is
unminimized or gets explicitly focused again. Only floating clients can
be minimized, so we move a tiled client from the 'tiling' to the
'floating' structures in a tag. Such a client still has 'floating' set
to false and will be moved back to the tiling layer if it gets
un-minimized.

There is no way to manipulate the minimization state through
EWMH-Requests, but I found the following keybinding helpful for testing.

    herbstclient keybind Alt-m attr clients.focus.minimized toggle