Implement client minimization (#1032)
This adds a new attribute 'minimized' that hides a client until it is
unminimized. Only floating clients can be minimized, so we move a tiled
client from the 'tiling' to the 'floating' structures in a tag. Such a
client still has 'floating' set to false and will be moved back to the
tiling layer if it gets un-minimized.

It is important that hlwm ensures that minimized clients can not be
focused. If a minimized client shall be focused, it has to be
unminimized first. This is not yet respected by the present PR in the
cycle_all command but must be implemented next.

There is no way to manipulate the minimization state through
EWMH-Requests, but I found the following keybinding helpful for testing.

    herbstclient keybind Alt-m attr clients.focus.minimized toggle