Fix moving a minimized client between tags
This sets the visibility of minimized clients correctly when they are
moved between tags. The test cases for it use the 'apply_rules' command
to move a client between tags since there is no direct via to do so via
the object system but it is possible via EWMH (see the test case
test_ewmh_move_client_to_tag), but I wanted the test cases in the
present change to be independent of X11.