Let cycle_all skip minimized clients (#1036)
Following the principle of #1032 that minimized clients shall never be
focused or visible, the cycle_all command must skip minimized clients.
This implements the skipping of minimized clients and adds a number of
test cases for it.

While at it, I've removed a unnecessary call to the raise() function of
the newly focused client. It was unnecessary because the raise()
function is called anyway at the end of cycleAllCommand() for floating
clients.