Extend ArgParse with boolean command line flags
This adds support for command line flags to the ArgParse class. A flag
can appear at any position in the command line arguments between the
(mandatory or optional) positional arguments. A flag is usually prefixed
with one or two dashes, but there is no formal requirement for this in
the ArgParse implementation.

A unknown flag will simply be ignored and parsed as a positional
argument instead. This is because I don't want negative numbers on the
command line to be misinterpreted as unknown command line flags.