New command 'mirror' for mirroring the frame tree (#1046)
This adapts the 'mirror' implementation from the PR #93 to the new
codebase. Additionally, it accepts a positional argument along which
axis the layout is to be mirrored.

On the implementation side, it uses type_traits and a new 'is_finite'
predicate to avoid boilerplate code when instantiating the Converter
class to the enum class type Horizontal/Vertical/Both. There is still a
bit to write, but much less compared to the Converter-implementation for
the ClientPlacement class.

In order to make Converter applicable to type traits, a new template
parameter to Converter "typename = void" was necessary.