Add basic python bindings
This initiates python bindings to herbstluftwm. The purpose of this is
to make it easier to write python scripts for hlwm. Moreover, as soon as
there are more useful functions in the bindings, we can also make use of
them in the test suite. Thus, we don't need a dedicated testsuite for
the bindings.

So far, the only functionality is call(), unchecked_call(), and
argument parsing.