Do not call applyLayout() from FrameLeaf (#1061)
In FrameLeaf::split(), the function FrameLeaf::setSelection() was called before
`selection` was (and could be) set correctly. Since setSelection() called
Monitor::applyLayout(), this led to a crash because the `selection` was out of
bounds.

The present change removes all invokations of Monitor::applyLayout() from
setSelection() (the only leftover use of applyLayout() in FrameLeaf/FrameSplit).
To ensure that the actual window geometries update correctly, I went
through all invocations of setSelection and added the applyLayout() where it
wasn't called before (namely, the commands focus_nth and cycle). The other
commands (cycle_all, remove, focus) called applyLayout() already.

Also, I removed the call `clients[selection]->window_focus()` because it is
simply wrong if the FrameLeaf is not focused. Instead, the applyLayout()
function should take care of the focus.

Beside adding a test case for the crash scenario of #1056, I've extended the
test cases for focus_nth and cycle to verify that focus indeed updates
correctly.

This fixes #1056.