Migrate CI to github actions (#1067)
This converts the travis ci configuration to github actions. Again we
have three jobs: one with gcc, one with clang, both on a current 64 bit
system, and one on an old 32 bit system.

All three jobs run in docker containers that are pulled from dockerhub.
In the old travis setup, if the docker file changed, then the newer
docker file was used directly. This is not yet reflected in the github
actions, so we can't use pull requests to test changes in the docker
files, yet.

Moreover, the ci/build.py is really optimized towards being invoked only
once, as it was the case for the travis ci. With github actions, we will
need to invoke it multiple times per job (namely step by step), so
ci/build.py will need to change further in the future. I tried to be as
little invasive here as possible to keep the diff of ci/build.py small.