Create new Floating namespace in floating.h
This moves all floating functions into a common namespace called
Floating. If the function had a prefix floating_ then I already
converted the name to camelCase, otherwise I left the name as it was.

The namespace Floating is implemented as a class with static functions
because in the future, these may become non-static by using member
variables of the Floating class instead of the global hlwm state (as it
is now when calling get_current_monitor() or focus_client()).