json doc: Derive writable property from source code (#1082)
This extends the doc/gendoc.py by a parser for the writable flag of
attributes. The parser looks for constructor arguments and invokations
of Attribute::setWritable(). Unfortunately, some of the latter were
indirect in loops, so I had to explicitly call setWritable() for every
attribute (happens in Client and Monitor). Since all attributes of the
Settings object are writable, I've hardcoded this in the gendoc.py.

Fortunately, we can be sure that we didn't miss a attribute, because the
tests in test_doc verify that the `writable` field in the json doc is
correct.