Add xft support for window titles
Use xft fonts in the window title rendering per default and only fall
back to plain x fonts if a XLFD is given explicitly.

Though xft supports utf8 per default, it does not seem to provide a font
fallback system if a particular unicode character is not contained in
the font selected. Hence, the best compromise here is to pick a font
supporting as many unicode characters as possible, e.g. Dejavu Sans.

This introduces a new dependencies on libxft and freetype which should
be standard on nowadays systems with X11.