Basic window title drawing (#1090)
The title of every client is drawn on the window decoration, if
title_height > 0. The drawing routine XDrawString does not support utf8
characters and only basic X fonts (no Xft). Both issues can be fixed in
the future.

This adds the new attribute type 'Font' and introduces new theme
attributes: title_font, title_height, title_color.

It is quite hard to add automated tests for it. Hence, for the time
being, there are only tests for the new font types.