Support proper utf8 strings in window titles
This uses the XFontSet structure and XmbDrawString routines that are
capable of utf8 encoded strings. The setlocale() in the main is of great
importance, because otherwise, text drawing stops on the first non-ascii
character.

Unfortunately, XCreateFontSet() or XmbDrawString() seem to be
incompatible with the terminus font on my installation.

If the XCreateFontSet fails, then we fall back to plain XFontStruct and
XDrawString.