Make the lock_tag attribute writable (#1100)
For some reason, the lock_tag attribute was not writable (even in the
old source code at tag v0.7.2). In fact, the lock_tag and unlock_tag
commands do nothing except setting the attribute.

While at it, I've added a bunch of test cases for the lock_tag attribute
and the respective commands.