Let objects have documentation (#1112)
This makes Object inherit from HasDocumentation and parses the
respective calls to setDoc() into the json document. The attr command
now prints this object documentation, and this allows us to add test
cases that it matches the doc in the json file.

Note that the object documentation is something different than the
documentation on these "child entries" because e.g. the documentation on
clients.focus is "the focused client" whereas the documentation on
Client is that it represents "a managed window".