Add attribute watchers (#1111)
This implements a commonly requested feature allowing to be notified
when an attribute changes. The implementation just saves the value of
the watched attributes and periodically checks if the value has been
changed.

There is still room for optimization: One optimization is to replace the
plain map from string to string to a more advanced structure that speeds
up scanForChanges(); another is to use the signals where possible.

This fixes #1098. Also, this implements the missing piece of #71.