Correctly interpret a panel's WM_STRUT (#1113)
The entries of WM_STRUT need to be interpreted with respect to the
borders of the root window (not of the monitor). The present change now
translates WM_STRUT into a rectangle which is then used to allocate
space at the monitor edge.

Only if WM_STRUT is empty, the geometry of the panel window is used as
the rectangle.

Formerly, we ignored WM_STRUT and always used the panel window, which
broke the support for panels that deliberately leave some space around
the panel. I originally took the window geometry, because 1. it was
easier (but as it turned out wronger) than computing the rectangle from
WM_STRUT and 2. this check for the case where the panel geometry and the
WM_STRUT were contradictory, as in #815.

This fixes #1110.