www: Generate object tree documentation (#1122)
This adds a script that transforms the json documentation into a
user-friendly asciidoc document. This asciidoc document is added to the
website (under "documentation" -> "objects"). The goal is that in the
future, we can use this generated document as a replacement for the
hand-written "objects" section in the man page.

For testing, run `make` in the www folder and open `object-doc.html`.