Use generated object doc in the man page
Use the asciidoc file generated by format-doc.py in herbstluftwm.txt. To
make it fit into the man page and to make the include mechanism work
also for the website, a few adjustments were necessary.

Unfortunately, if a include fails, then asciidoc does not fail but only
prints a warning.