cmake: Fix dependencies between asciidoc files
In the current build system, herbstluftwm.html and herbstluftwm.1 were
not rebuilt after hlwm-objects-gen.txt had been changed. The issue is
that this file was only a dependency of the target but not of the
command. The present change makes it a dependency of the command by
passin it as an optional argument to gen_manpage and gen_html
(add_dependencies does not work for extending add_custom_commands, only
for add_custom_target).