Fix frame objects in 'mirror' and 'rotate'
The commands 'mirror' and 'rotate' updated the child frame pointers
explicitly without updating the pointers for the frame objects. Which
caused the frame object tree to be wrong.

The present change fixes this and adds a test for it.

Thanks to @astier for an excellent bug report on this!

Fixes #1151.