Add test cases for decorations and fonts
This adds a screenshot-based test suite for the drawing of decorations.
The idea is that after changing the theme or other parameters of the
window decorations, we take a screenshot and tell it from the pixels
whether everything is drawn correctly.

So far, this contains tests for the main, inner, and outer border color
and tests for the font-rendering.

Also this, adds another x11-sync to x11's sync_with_hlwm(), because
syncing with hlwm itself only makes sense if all the client information
is already present at the xserver.