Build doc in CI
This creates a new cmake target for the doc and calls it another ci
build job. While at it, I've fixed a race condition in our cmake setup,
which was caused because the json-target was used before gen_json_doc()
created it (thus the order change in doc/CMakeLists.txt).