Do not rebuild doc in tarball builds
In source tarballs, the built documentation is already shipped along
with the tarball. The CMakeLists of doc now detects a prebuilt
documentation, and in this case simply copies the documentation to the
build directory (without requiring asciidoc to be installed).

In the CI job for the docs, the tarball is now built.