Re-use pre-built doc in tarball builds (#1171)
In source tarballs, the built documentation is already shipped along
with the tarball. The CMakeLists of doc now detects a prebuilt
documentation, and in this case simply copies the documentation to the
build directory (without requiring asciidoc to be installed).

This implies that one does not need to (and in fact should not) disable
`WITH_DOCUMENTATION` when building a tarball without having asciidoc
installed.

In the CI job for the docs, the tarball is now built and tested. In the
ancient ubuntu build, this tarball is used instead of the source from
the git repo. To be sure that everything worked, the job verifies that
the man page has been installed to the DESTDIR.