Make Colors rgba
Extend Color by a alpha value and extend the parser and printer to print
colors in the #rrggbbaa format. The alpha value is not used anywhere at
the moment, but it will come in handy for decoration drawing.

Also, this adds a new test suite file dedicated to all the type parser
and printers.