Initialize rectangle before XGetGeometry()
If a window disappears right before XGetGeometry() is called, then the
rectangle is not initialized. Hence, just initialize everything before
to be on the save side.

This was detected by address sanitization, reporting that jumps depended
on uninitialized memory, and it was triggered by the qt application
katarakt that creates two windows and hides them instantly again.