Make default colors more transparent and quiet
Stop using the #9fbc00 color as the default border color, since it is a
quite aggressive color. I've always loved this color, when it was
introduced by pwmt.org applications, but I've stoped using it in my own
settings some years ago.

Instead, the border color now is the same (dark) green tone as in the
frame border, and is optimized such that it also looks nice with or
without transparency.