Support true transparency in all decorations
This adds true transparency support in all frame and client decorations
(using the xrender X11 extension), where the level of transparency is
controlled by the alpha value of the colors. The only instance where the
alpha value does not work yet is at Xft fonts.

Per default, hlwm tries to initialize true transparency on startup (this
can be prevented with --no-transparency).

The test cases for the decoration colors now run in all combinations of
xrender being available at the server and hlwm trying to use it.
For xrender to work, we need to start xvfb with color depth 24 (rather
than 8), and this also results in the screenshots being in a slightly
different format.

On the technical side, the XConnection class now provides a singleton
object getter, which can be used to obtain the pointers to Visual and
Colormap. These structures are either the 'default' or specia structures
if true transparency is used.

The new library dependency xrender should be available on all systems,
but we can think of making it an optional dependency, since
XConnection::tryInitTransparency() is the only function using the
library.