Make default colors more transparent and quiet (#1185)
Stop using the #9fbc00 color as the default border color, since it is a
quite aggressive color. I loved this color, when it was introduced by
pwmt.org applications, but I've stopped using it in my own settings some
years ago and I have the impression that many people don't like it when
starting with hlwm (and the default hlwm config).

Instead, the border color now is the same (dark) green tone as in the
frame border, and is optimized such that it also looks nice with or
without transparency.