Cut holes into frames behind tiled clients
For semi-transparent clients, it is unpleasant to see the frame
decoration shining through. Hence, we now cut holes into the frame's
decoration at the area where clients are. If no true transparency is
activated, no difference should be noticable, and for true transparency,
one sees the wallpaper or desktop window shining through.

Internally, the last 'outer' geometry of every client is used, and so
this also interacts nicely with pseudotiled windows and the grid layout.

This should improve the cosmetics and implicitly also solve #1183.