Add converter for Monitor*
This introduces a new converter class for the parsing of descriptions of
objects, e.g. Monitors, Tags, Clients. So far, only the converter for
Monitor* is implemented. If a command gets a monitor as a parameter, we
can now use ArgParse which internally uses the Converter class.

This new converter is used in the commands remove_monitor and tag_status
and their test cases have slightly been adjusted and extended.