Run-Time Converter for Client*
This implements RunTimeConverter for Client* and uses it in the commands
jumpto and bring which are moved to GlobalCommands.

Since there are multiple variables and functions are called 'complete',
RunTimeConverter::complete is renamed to RunTimeConverter::completeEntries
to make the variable names a bit more unique.

The test cases with error handling are slightly extended and moved to
test_global_commands.py, but the existing 'positive' test cases for
bring and jumpto are still in test_clients and test_layout in order to
keep the diff small.