Integrate completion into ArgParse
This introduces the new type CallOrComplete that combines the invocation
and argument completion of a command.

To illustrate the new functionality, it is applied to the commands
bring, focus, shift, tag_status, remove_monitor whose test suite
regarding parse errors and completion is respectively extended.

The word-list suggestion parameter is not yet used, but it will be
useful for commands like cycle_all (completion for -1 and +1) or resize
(completion for +0.02 and -0.02), which will be handled in separate
commits.