Make frame attributes writable
The frame attributes of FrameLeaf and FrameSplit are now writable and
allow to adjust the frame locally. As usual, the test cases consume more
lines than the actual source code accomplishing this change.

In the source code, the getTag() function seemed to be unused.
Unfortunately, I had to remove the 'const' qualifier from getFraction()
because otherwise the DynAttribute_ constructor failed to match.

This is another step towards exposing the frame tree via objects (#2).