Create EWMH object long before Root object (#1197)
For the implementation of #838, it is beneficial to have the EWMH object
set up much earlier during bootup: 1. in order to have the WM window
available much earlier and 2. to implement the window manager
replacement in the EWMH object where it belongs.

A few new tests verify that the EWMH object is still set up correctly.