Option --replace to replace existing window manager
Herbstluftwm now uses the ICCCM-speicified way of checking whether there
is an existing window manager, namely by checking the ownership
particular 'selection'. The owner of this selection is the active window
manager. This change introduces the new features:

    1. herbstluftwm does not start if the selection is already owned
    2. if herbstluftwm is started with --replace, try to become the
       owner of the selection
    3. if herbstluftwm looses the ownership, then quit, because there is
       some other window manager booting up.

In the test case, we only check that herbstluftwm can replace itself,
but I've tested it locally and the replacement works (forth and back)
with xfwm4, openbox, fvwm2, pekwm.

This change also adds a test case for the 'quit' command because it is
essentially half the test of '--replace'.

This implements #838.