Manage desktop windows in global stack
Instead of lowering desktop windows when they appear, manage desktop
windows in the global stack explicitly. Unfortunately, I did not find a
way to query the window stack. So for debugging and testing the output
of the `stack` command is extended by the desktop windows' winid and
window instance/class.

This solves an issue here where the Xfce's desktop window covered all
normal windows if I run `herbstluftwm --replace` in a running Xfce
session.