Migrate 'raise' command to GlobalCommands and ArgParse (#1211)
The raise command expects as the parameter either a client or the window
id of an unmanaged window. This is reflected by the type
Either<Client*,WindowID> for the new template class Either. Because of
the genericity of the Converter of Either, the error message might be a
bit cryptic.