Objects for panels (#1214)
This makes PanelManager an object containing children for all panels. A
panel's window attribute is the first attribute to use the type WindowID
(whose converter has been used before already).

The documentation of a class is only included in the man page if there
is a path that leads to such an object. For panels, we can't provide
such a path in general. As a fix, instead of the 'name' of an entry in
an object, the json doc now mentions a 'name_pattern'. This
'name_pattern' can be put into the documentation, but the test cases in
test_doc.py must not treat this name literally.