New objects for types documentation
This extends the object tree to also contain meta-information on the
available types. In order to keep the diff small, only the most basic
types are included.

The table defining the names of the types has not been used really (only
in the 'help' command so far). To make the names more uniform, I have
adjusted the names such that they match to the type argument of the
`new_attr` command (in the long run, `new_attr` should of course work
for all entries of the Types enum).

Since the objects essentially consist of documentation, there is not
much to test.