Migrate use, use_index to ArgParse
This makes use of ArgParse in the commands 'use' and 'use_index', and
makes them part of the GlobalCommands class. They both need access to
TagManager and MonitorManager, so they don't really belong to neither of
those classes. Since they modify the global state, they fit quite well
in GlobalCommands.

For the implementation, TagManager now provides a run-time converter
for tags, but it only can be used for 'use'.

Also, this adds test cases for both commands and thus also for the
run-time converter.