Use posix escaping in Completion objects
In the objects of the Completion class, the shellOutput_ member was not
used at all, and so the completion results were not escaped correctly.

Now, the results are escaped and while at it, the posix_escape()
function is converted to C++. This makes the implementation much simpler
and it even fixes a memory leak in command.cpp (prefix_escaped is never
free'd).

A new special case in the posix_escape() function is the empty string,
which is now escaped correctly (before, zsh just added a space to the
command line, which of course did not have any effect).

In the tests, there is a new case for the empty string, and the escaping
of the other characters is now tested for the commands 'use' and 'move'
which still use the old command completion but will soon be migrated
(for 'use', see #1224).